fix: removed weird types in lib. feat: gitignore: more tmp files

+++ repo/include/my.h
@@ -62,7 +62,7 @@ char **my_str_to_word_array(char const *str, char sep, char skip);
 
 // my_str
 
-size_t my_strlen(char const *str);
+int my_strlen(char const *str);
 char *my_strcat(char *dest, char const *src);
 int my_strcmp(char const *s1, char const *s2);
 char *my_strdup(char const *src);
